Assignment 2/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
PyYAML>=6.0
python-dotenv>=1.0.0